+++ tests/unit/Karla/KarlaTest.php
@@ -66,15 +66,42 @@ class KarlaTest extends PHPUnit\Framework\TestCase
     }
 
     /**
-     * Test
+     * Test that the constructor is public and creates a valid Karla instance.
      *
      * @test
      *
      * 
      */
-    public function privateMethods()
+    public function constructor()
     {
-        $this->assertInstanceOf('Karla\Karla', Karla::perform(PATH_TO_IMAGEMAGICK));
+        $this->assertInstanceOf('Karla\Karla', new Karla(PATH_TO_IMAGEMAGICK));
+    }
+
+    /**
+     * Test that the constructor throws an exception for an invalid path.
+     *
+     * @test
+     *
+     * 
+     */
+    public function constructorWithInvalidPath()
+    {
+        $this->expectException(\InvalidArgumentException::class);
+        new Karla('/going/nowhere/');
+    }
+
+    /**
+     * Test that each call to perform() returns a fresh instance.
+     *
+     * @test
+     *
+     * 
+     */
+    public function performReturnsFreshInstance()
+    {
+        $first  = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $second = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $this->assertNotSame($first, $second);
     }
 
     /**
@@ -86,7 +113,7 @@ class KarlaTest extends PHPUnit\Framework\TestCase
      */
     public function raw()
     {
-        $karla = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $karla = new Karla(PATH_TO_IMAGEMAGICK);
         $this->assertTrue(preg_match('/Version\:\sImageMagick/', $karla->raw('identify', '--version')) == 1);
     }
 
@@ -99,7 +126,7 @@ class KarlaTest extends PHPUnit\Framework\TestCase
      */
     public function convert()
     {
-        $karla = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $karla = new Karla(PATH_TO_IMAGEMAGICK);
         $this->assertInstanceOf('Karla\Program\Convert', $karla->convert());
     }
 
@@ -112,7 +139,7 @@ class KarlaTest extends PHPUnit\Framework\TestCase
      */
     public function identify()
     {
-        $karla = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $karla = new Karla(PATH_TO_IMAGEMAGICK);
         $this->assertInstanceOf('Karla\Program\Identify', $karla->identify());
     }
 
@@ -125,7 +152,7 @@ class KarlaTest extends PHPUnit\Framework\TestCase
      */
     public function composite()
     {
-        $karla = Karla::perform(PATH_TO_IMAGEMAGICK);
+        $karla = new Karla(PATH_TO_IMAGEMAGICK);
         $this->assertInstanceOf('Karla\Program\Composite', $karla->composite());
     }
 }